Moved geodesy to geometry
